feat: extract signals from the closest acceleromenter to the fault source
exemple: if the fail is in a bearing in the Drive end (DE), the vibration signal will be extracted from the acceleromenter placed in DE. If tha fail is in the Fan end (FE), the vibration signal will be extracted from the acceleromenter placed in FE.

Normal acquisitions are extracted twice. There are samples from both, drive end adn fan end.